+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [0.0.2] - 2024-05-03
+### Fixed
+- gradient calculation if both values are in the same color range
+### Added
+- recommended temperature colors export
+### Breaking
+- TypeScript `lib` field updated to `es2023`
+
 ## [0.0.1] - 2024-04-25
 ### Added
 - first version